+++ src/main/java/com/silverminer/shrines/config/DefaultStructureConfig.java
@@ -3,10 +3,16 @@ package com.silverminer.shrines.config;
 import com.google.common.collect.Lists;
 import com.silverminer.shrines.structures.load.DefaultedStructureData;
 
+import net.minecraft.world.biome.Biome.Category;
+
 public class DefaultStructureConfig {
+	public static final DefaultedStructureData CUSTOM = new DefaultedStructureData("", "", 0).setDistance(50)
+			.setSeperation(8);
+
 	public static final DefaultedStructureData ABANDONEDWITCHHOUSE_CONFIG = new DefaultedStructureData(
-			"Abandoned Witch House", "abandoned_witch_house", 1721882513)// .setBiomes(Category.SWAMP, Category.FOREST)
-					.addToBiomeBlacklist("minecraft:flower_forest", "minecraft:tall_birch_forest", "minecraft:forest",
+			"Abandoned Witch House", "abandoned_witch_house", 1721882513)
+					.setBiomeCategoryWhitelist(Category.SWAMP, Category.FOREST)
+					.setBiomeBlacklist("minecraft:flower_forest", "minecraft:tall_birch_forest", "minecraft:forest",
 							"minecraft:birch_forest", "minecraft:birch_forest_hills")
 					.setDistance(60).setSeperation(11);
 
@@ -17,37 +23,36 @@ public class DefaultStructureConfig {
 			.setDistance(70).setSeperation(12).setUseRandomVarianting(false);
 
 	public static final DefaultedStructureData ENDTEMPLE_CONFIG = new DefaultedStructureData("End Temple", "end_temple",
-			32 ^ 478392).setDistance(60).setSeperation(11)// .setBiomes(Category.THEEND)
-					.addToBiomeBlacklist("minecraft:the_end", "minecraft:the_void", "minecraft:small_end_islands")
+			32 ^ 478392).setDistance(60).setSeperation(11).setBiomeCategoryWhitelist(Category.THEEND)
+					.setBiomeBlacklist("minecraft:the_end", "minecraft:the_void", "minecraft:small_end_islands")
 					.setDimensionWhitelist(Lists.newArrayList("minecraft:the_end"));
 
 	public static final DefaultedStructureData FLOODEDTEMPLE_CONFIG = new DefaultedStructureData("Flooded Temple",
-			"flooded_temple", 54315143).setDistance(50).setSeperation(10)
-					.setUseRandomVarianting(false)/*
-													 * .removeBiome(Category.DESERT) .removeBiome(Category.MESA)
-													 */;
+			"flooded_temple", 54315143).setDistance(50).setSeperation(10).setUseRandomVarianting(false)
+					.removeFromBiomeCategoryWhitelist(Category.DESERT, Category.MESA);
 
 	public static final DefaultedStructureData GUARDIANMEETING_CONFIG = new DefaultedStructureData("Guardian Meeting",
 			"guardian_meeting", 1498473232).setDistance(70).setSeperation(17).setUseRandomVarianting(false);
 
 	public static final DefaultedStructureData HARBOUR_CONFIG = new DefaultedStructureData("Harbour", "harbour",
-			651398043).setDistance(50).setSeperation(8);// .setBiomes(Biome.Category.OCEAN);
+			651398043).setDistance(50).setSeperation(8).setBiomeCategoryWhitelist(Category.OCEAN);
 
 	public static final DefaultedStructureData HIGHTEMPLE_CONFIG = new DefaultedStructureData("High Tempel",
 			"high_temple", 536987987).setDistance(85).setSeperation(18);
 
 	public static final DefaultedStructureData JUNGLETOWER_CONFIG = new DefaultedStructureData("Jungle Tower",
-			"jungle_tower", 987531843).setDistance(60).setSeperation(11);// .setBiomes(Category.JUNGLE);
+			"jungle_tower", 987531843).setDistance(60).setSeperation(11).setBiomeCategoryWhitelist(Category.JUNGLE);
 
 	public static final DefaultedStructureData MINERALTEMPLE_CONFIG = new DefaultedStructureData("Mineral Temple",
 			"mineral_temple", 576143753).setDistance(50).setSeperation(10).setUseRandomVarianting(false);
 
 	public static final DefaultedStructureData NETHERPYRAMID_CONFIG = new DefaultedStructureData("Nether Pyramid",
 			"nether_pyramid", 7428394).setDistance(150).setSeperation(50)
-					.addDimensionToWhitelist("minecraft:the_nether");// .addBiome(Category.NETHER);
+					.addDimensionToWhitelist("minecraft:the_nether").addToBiomeCategoryWhitelist(Category.NETHER);
 
 	public static final DefaultedStructureData NETHERSHRINE_CONFIG = new DefaultedStructureData("Nether Shrine",
-			"nether_shrine", 653267).setDistance(80).setSeperation(15).addDimensionToWhitelist("minecraft:the_nether");// .addBiome(Category.NETHER);
+			"nether_shrine", 653267).setDistance(80).setSeperation(15).addDimensionToWhitelist("minecraft:the_nether")
+					.addToBiomeCategoryWhitelist(Category.NETHER);
 
 	public static final DefaultedStructureData ORIENTALSANCTUARY_CONFIG = new DefaultedStructureData(
 			"Oriental Sanctuary", "oriental_sanctuary", 143665).setDistance(50).setSeperation(14);