+++ src/main/java/com/silverminer/shrines/gui/packets/StructuresPacketsScreen.java
@@ -35,6 +35,11 @@ public class StructuresPacketsScreen extends Screen {
 		this.packets = packets;
 	}
 
+	public void onClose() {
+		this.minecraft.setScreen(this.lastScreen);
+		ShrinesPacketHandler.sendToServer(new CTSPlayerLeftQueuePacket(this.minecraft.player.getUUID()));
+	}
+
 	protected void init() {
 		this.minecraft.keyboardHandler.setSendRepeatsToGui(true); // Why should I have this?
 		this.searchBox = new TextFieldWidget(this.font, (this.width / 4) * 3, 3, 100, 20, this.searchBox,
@@ -50,8 +55,7 @@ public class StructuresPacketsScreen extends Screen {
 				}, this.packets);
 		this.addButton(new ImageButton(2, 2, 91, 20, 0, 0, 20,
 				new ResourceLocation(ShrinesMod.MODID, "textures/gui/widgets.png"), 256, 256, (button) -> {
-					this.minecraft.setScreen(this.lastScreen);
-					ShrinesPacketHandler.sendToServer(new CTSPlayerLeftQueuePacket(this.minecraft.player.getUUID()));
+					this.onClose();
 				}, StringTextComponent.EMPTY));
 		this.delete = this.addButton(new Button(this.width / 2 - 80 - 80 - 9, this.height - 22, 80, 20,
 				new TranslationTextComponent("Delete"), (button) -> {
@@ -67,7 +71,8 @@ public class StructuresPacketsScreen extends Screen {
 				}));// TRANSLATION
 		this.rename = this.addButton(new Button(this.width / 2 + 80 + 9, this.height - 22, 80, 20,
 				new TranslationTextComponent("Rename"), (button) -> {
-					this.list.getSelectedOpt().ifPresent(StructurePacketsList.Entry::rename);
+					this.list.getSelectedOpt().ifPresent(entry -> this.minecraft
+							.setScreen(new RenameStructurePacketScreen(this, entry.getPacket())));
 				}));// TRANSLATION
 		this.updateButtonStatus(false);
 		this.children.add(this.searchBox);