+++ src/main/java/com/silverminer/shrines/utils/network/cts/CTSFetchStructuresPacket.java
@@ -42,9 +42,7 @@ public class CTSFetchStructuresPacket implements IPacket {
 	}
 
 	public static void handle(CTSFetchStructuresPacket packet, Supplier<NetworkEvent.Context> context) {
-		context.get().enqueueWork(() -> {
-			Handle.handle(packet);
-		});
+		context.get().enqueueWork(Handle.handle(packet));
 		context.get().setPacketHandled(true);
 	}