+++ src/main/java/com/silverminer/shrines/utils/network/cts/CTSPlayerLeftQueuePacket.java
@@ -30,9 +30,7 @@ public class CTSPlayerLeftQueuePacket implements IPacket {
 	}
 
 	public static void handle(CTSPlayerLeftQueuePacket packet, Supplier<NetworkEvent.Context> context) {
-		context.get().enqueueWork(() -> {
-			Handle.handle(packet);
-		});
+		context.get().enqueueWork(Handle.handle(packet));
 		context.get().setPacketHandled(true);
 	}